move extension to their related extension class
