--- src/blueprints/Do.Blueprints.Service.Application/Extensions/GuidExtensions.cs
@@ -1,17 +1,0 @@
-using Do.Testing;
-
-namespace Do;
-
-public static class GuidExtensions
-{
-    public static Guid AGuid(this Stubber _,
-        string? starts = default
-    )
-    {
-        starts ??= string.Empty;
-
-        const string template = "4d13bbe0-07a4-4b64-9d31-8fef958fbef1";
-
-        return Guid.Parse($"{starts}{template[starts.Length..]}");
-    }
-}
