--- src/blueprints/Do.Blueprints.Service.Application/Extensions/IntegerExtensions.cs
@@ -1,8 +1,0 @@
-using Do.Testing;
-
-namespace Do;
-
-public static class IntegerExtensions
-{
-    public static int AnInteger(this Stubber _) => 42;
-}
