--- src/blueprints/Do.Blueprints.Service.Application/Extensions/StringExtensions.cs
@@ -1,14 +1,0 @@
-using Do.Testing;
-
-namespace Do;
-
-public static class StringExtensions
-{
-    public static string AnEmail(this Stubber _) => "info@test.com";
-
-    public static string AString(this Stubber _,
-        string? value = default
-    ) => value ?? "test string";
-
-    public static Guid ToGuid(this string source) => Guid.Parse(source);
-}
