--- src/blueprints/Do.Blueprints.Service.Application/Extensions/TimeProviderExtensions.cs
@@ -1,33 +1,0 @@
-using Do.Core.Mock;
-using Do.Testing;
-
-namespace Do;
-
-public static class TimeProviderExtensions
-{
-    public static TimeProvider TheTime(this Mocker mockMe,
-        DateTime? now = default,
-        bool passSomeTime = false,
-        bool reset = false
-    )
-    {
-        var fakeTimeProvider = (ResettableFakeTimeProvider)mockMe.Spec.GiveMe.The<TimeProvider>();
-
-        if (reset)
-        {
-            fakeTimeProvider.Reset();
-        }
-
-        if (now is not null)
-        {
-            fakeTimeProvider.SetUtcNow(new(now.Value, fakeTimeProvider.LocalTimeZone.BaseUtcOffset));
-        }
-
-        if (passSomeTime)
-        {
-            fakeTimeProvider.Advance(TimeSpan.FromSeconds(1));
-        }
-
-        return fakeTimeProvider;
-    }
-}
